build cleanup
